+++ solutions/p15/p15.mojo
@@ -21,7 +21,7 @@ fn axis_sum[
 ](
     output: LayoutTensor[dtype, out_layout, MutAnyOrigin],
     a: LayoutTensor[dtype, in_layout, ImmutAnyOrigin],
-    size: Int,
+    size: UInt,
 ):
     global_i = block_dim.x * block_idx.x + thread_idx.x
     local_i = thread_idx.x
@@ -50,7 +50,7 @@ fn axis_sum[
     barrier()
 
     # do reduction sum per each block
-    stride = TPB // 2
+    stride = UInt(TPB // 2)
     while stride > 0:
         # Read phase: all threads read the values they need first to avoid race conditions
         var temp_val: output.element_type = 0
@@ -92,7 +92,7 @@ def main():
         ctx.enqueue_function_checked[kernel, kernel](
             out_tensor,
             inp_tensor,
-            SIZE,
+            UInt(SIZE),
             grid_dim=BLOCKS_PER_GRID,
             block_dim=THREADS_PER_BLOCK,
         )
